# Optimization vector add