--- a/oneflow/core/vectorized/vec.h
+++ b/oneflow/core/vectorized/vec.h
@@ -28,9 +28,6 @@ namespace oneflow {
 template<typename T, typename Enable = void>
 class VecFunc;
 
-
-void tmp_test();
-
 template<typename T>
 class VecFunc<T, typename std::enable_if<std::is_same<T, float>::value
                                          || std::is_same<T, double>::value>::type> {
@@ -46,12 +43,6 @@ class VecFunc<T, typename std::enable_if<std::is_same<T, float>::value
         sub_func = VectorizedDefault<T>::sub;
         mul_func = VectorizedDefault<T>::mul;
         div_func = VectorizedDefault<T>::div;
-        // test_func = tmp_test;
-        // auto v_p = *(test_func.target<void(*)()>());
-        // printf("v_p is %p\n", v_p);
-        // test_func();
-        // std::cout <<"test_func address = " << std::addressof(test_func) << std::endl;
-        // printf("VectorizedDefault add address = %p \n", VectorizedDefault<T>::add);
         break;
       }
 #ifdef WITH_AVX
@@ -97,34 +88,33 @@ class VecFunc<
     auto capability = compute_cpu_capability();
     switch (capability) {
       case CPUCapability::DEFAULT:
+      {
         std::cout << "DEFAULT : NO SIMD  int" << std::endl;
         add_func = VectorizedDefault<T>::add;
         sub_func = VectorizedDefault<T>::sub;
         mul_func = VectorizedDefault<T>::mul;
         div_func = VectorizedDefault<T>::div;
-        // test_func = tmp_test;
-        // std::cout <<"test_func address = " << std::addressof(test_func) << std::endl;
-        // printf("VectorizedDefault add address = %p \n", VectorizedDefault<T>::add);
-
-        // auto vp = *(test_func.target<void(*)()>());
-        // printf("vp is %p\n", vp);
-
         break;
+      }
 #ifdef WITH_AVX
       case CPUCapability::AVX2:
+      {
         std::cout << "AVX2" << std::endl;
         add_func = VectorizedAvx2<T>::add;
         sub_func = VectorizedAvx2<T>::sub;
         mul_func = VectorizedAvx2<T>::mul;
         div_func = VectorizedAvx2<T>::div;
         break;
+      }
       case CPUCapability::AVX512:
+      {
         std::cout << "AVX512" << std::endl;
         add_func = VectorizedAvx512<T>::add;
         sub_func = VectorizedAvx512<T>::sub;
         mul_func = VectorizedAvx512<T>::mul;
         div_func = VectorizedAvx512<T>::div;
         break;
+      }
 #endif
       default: break;
     }
@@ -135,7 +125,6 @@ class VecFunc<
   static std::function<void(size_t, size_t, T*, T*, T*)> sub_func;
   static std::function<void(size_t, size_t, T*, T*, T*)> mul_func;
   static std::function<void(size_t, size_t, T*, T*, T*)> div_func;
-  static std::function<void(void)> test_func;
 };
 
 void vectorized_init();