1 根据CPU类型选择运算方式，还可以通过手动设定。
```
 ONEFLOW_CPU_CAPABILITY=avx512     python ***.py
 ONEFLOW_CPU_CAPABILITY=avx2       python ***.py
 ONEFLOW_CPU_CAPABILITY=default    python ***.py
```

2 根据数据量分配计算线程

3 添加cpuinfo第三方库

4 add_n 耗时对比
```
测试方式
len = 30*224*224
a = flow.rand(len, dtype=flow.float)
b = flow.rand(len, dtype=flow.float)
flow.add(a, b)
以下测试结果已经排除Cache命中的影响。
多次循环取中间值结果：

当前oneflow commit ： 21caffd9d94e70538a9035abdcf215405d045167
____________________________________
add_n kernel 2760 us 
add_n kernel 2801 us 
add_n kernel 2701 us 
add_n kernel 2789 us 
add_n kernel 6143 us 
____________________________________
add_n kernel 2758 us 
add_n kernel 2798 us 
add_n kernel 2749 us 
add_n kernel 2764 us 
add_n kernel 2276 us 

此 PR 结果
____________________________________
add_n kernel 965 us 
add_n kernel 877 us 
add_n kernel 865 us 
add_n kernel 1119 us 
add_n kernel 925 us 
____________________________________
add_n kernel 893 us 
add_n kernel 872 us 
add_n kernel 886 us 
add_n kernel 960 us 
add_n kernel 872 us 
```
上述带来的速度度提升主要是线程池带来的影响，在开O3的情况下编译器会默认使用simd指令。运算的瓶颈来自内存。